Organized the tables router.
